hyfed-server/requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
PyJWT==1.7.1

